Refactor booking changed domain event describer
Extract logic from the `DomainEventDescriber` class and put it into a standalone `LegacyTimelineFactory`. This is in preparation for making changes to the logic.
